Export plugin, but also expose class for customization. + fix some errors
